Silence compiler warning in Picture.cpp
Warning was:

xbmc/pictures/Picture.cpp: In static member function 'static bool CPicture::CreateThumbnailFromSurface(const unsigned char*, int, int, int, const string&)':
xbmc/pictures/Picture.cpp:93:87: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   const bool ret = file.OpenForWrite(thumbFile, true) && file.Write(thumb, thumbsize) == thumbsize;
                                                                                       ^

+++ xbmc/pictures/Picture.cpp
@@ -76,7 +76,9 @@ bool CPicture::CreateThumbnailFromSurface(const unsigned char *buffer, int width
   }
 
   XFILE::CFile file;
-  const bool ret = file.OpenForWrite(thumbFile, true) && file.Write(thumb, thumbsize) == thumbsize;
+  const bool ret = file.OpenForWrite(thumbFile, true) &&
+                   file.Write(thumb, thumbsize) == static_cast<ssize_t>(thumbsize);
+
   pImage->ReleaseThumbnailBuffer();
   delete pImage;
